move read_oxide_barcode to packrat-vpd-loader
