+++ drv/local-vpd/Cargo.toml
@@ -9,7 +9,6 @@ zerocopy = { workspace = true }
 
 drv-i2c-api = { path = "../../drv/i2c-api" }
 drv-i2c-devices = { path = "../../drv/i2c-devices" }
-oxide-barcode = { path = "../../lib/oxide-barcode" }
 userlib = { path = "../../sys/userlib" }
 ringbuf = { path = "../../lib/ringbuf" }
